backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0

# AI/ML
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Optional: LLM Integration
openai>=1.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
